\name{R2WinBUGS-package}
\alias{R2WinBUGS-package}
\alias{R2WinBUGS}

\docType{package}

\title{Running WinBUGS and OpenBUGS from R / S-PLUS}

\description{

\pkg{R2WinBUGS} package provides possiblity to call a \pkg{BUGS} model,
summarize inferences and convergence in a table and graph, and save the
simulations in arrays for easy access in \R / S-PLUS. In S-PLUS, the
\pkg{OpenBUGS} functionality and the windows emulation functionality is
not yet available. The main command is \code{\link{bugs}}.

}

\details{

The following are sources of information on \pkg{R2WinBUGS} package:
\tabular{ll}{
  DESCRIPTION file\tab \code{library(help="R2WinBUGS")}\cr
  \tab \cr
  This file\tab        \code{package?R2WinBUGS}\cr
  \tab \cr
  Vignette\tab         \code{vignette("R2WinBUGS")}\cr
  \tab \cr
  Some help files\tab  \code{\link{bugs}}\cr
  \tab                 \code{\link{write.model}}\cr
  \tab                 \code{\link{print.bugs}}\cr
  \tab                 \code{\link{plot.bugs}}\cr
  \tab \cr
  News\tab             \code{file.show(system.file("NEWS", package="R2WinBUGS"))}\cr
}

}

\keyword{package}
